StreamlitApp/requirements.txt
streamlit
pandas
python-dotenv
requests
boto3
beautifulsoup4
langchain
langchain-openai
langchain-community
langgraph
lxml
chromadb
matplotlib